Refactor creators and fixed hierarchical graph
