+++ src/anemoi/graphs/__init__.py
@@ -7,4 +7,5 @@
 
 from ._version import __version__
 
+
 EARTH_RADIUS = 6371.0  # km
