+++ src/anemoi/graphs/creators/__init__.py
@@ -1,0 +1,4 @@
+from .create import GraphCreator
+from .create import HierarchicalGraphCreator
+
+__all__ = ["GraphCreator", "HierarchicalGraphCreator"]
